# api/hqbase-mail-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Webmail Mail API",
    "version": "1.0.0",
    "description": "Stable v1 mail API for Webmail clients, automations, and agents. Additive fields may be introduced within v1; clients must ignore unknown response fields. Administrative APIs are not part of this contract."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://mail.example.com",
          "description": "Webmail origin"
        }
      }
    }
  ],
...
  "paths": {
    "/api/v1/mailboxes": {
      "get": {
        "summary": "List mailboxes",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Mailboxes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mailbox"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Mailboxes"],
        "operationId": "listMailboxes"
      }
    },
...
    "/api/v1/labels": {
      "get": {
        "summary": "List labels",
        "description": "Lists shared workspace labels. Labels organize mail but never grant mailbox access or change folders.",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace labels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Label"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The mail:read permission is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Labels"],
        "operationId": "listLabels"
      }
    },
    "/api/v1/messages": {
      "get": {
        "summary": "List or search messages",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "One page of messages, newest activity first.",
            "headers": {
              "Link": {
                "required": false,
                "description": "RFC 8288 link to the next page, as `<url>; rel=\"next\"`. The header is absent on the last page.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_CURSOR` when the cursor is malformed or is not a message cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "listMessages",
        "parameters": [
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"]
            },
            "description": "Message folder to list. The `drafts` value remains for v1 compatibility, but current write paths do not store drafts as message rows. Use `/api/v1/drafts` for drafts."
          },
          {
            "in": "query",
            "name": "mailboxId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to one accessible mailbox."
          },
          {
            "in": "query",
            "name": "labelId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to messages assigned this existing label."
          },
          {
            "in": "query",
            "name": "labelIds",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "uniqueItems": true,
              "items": {
                "type": "string",
                "maxLength": 100
              }
            },
            "description": "Repeat this parameter to require every listed label on each returned message."
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Search one literal substring in the subject, sender, To recipients, snippet, and plain-text body. The characters %, _, and \\ are ordinary text."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of messages in the page."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page. Read it from the Link header and send it unchanged."
          },
          {
            "name": "includeLabels",
            "in": "query",
            "required": false,
            "description": "Include label membership. Does not filter results or change the cursor.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ]
      }
    },
...
    "/api/v1/conversations": {
      "get": {
        "summary": "List or search conversations",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Conversations"],
        "operationId": "listConversations",
        "parameters": [
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["inbox", "sent", "starred", "archived", "trash", "catchall"]
            },
            "description": "Folder to list."
          },
          {
            "in": "query",
            "name": "mailboxId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to one accessible mailbox."
          },
          {
            "in": "query",
            "name": "labelId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to conversations where an accessible message has this existing label."
          },
          {
            "in": "query",
            "name": "labelIds",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "uniqueItems": true,
              "items": {
                "type": "string",
                "maxLength": 100
              }
            },
            "description": "Repeat this parameter to require every listed label on accessible messages in each returned conversation."
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Search one literal substring in the subject, sender, To recipients, snippet, and plain-text body. The characters %, _, and \\ are ordinary text."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page."
          },
          {
            "name": "includeLabels",
            "in": "query",
            "required": false,
            "description": "Include label membership. Does not filter results or change the cursor.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ]
      }
    },
...
    "/api/v1/drafts": {
      "get": {
        "summary": "List drafts",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "One page of drafts, newest update first.",
            "headers": {
              "Link": {
                "required": false,
                "description": "RFC 8288 link to the next page, as `<url>; rel=\"next\"`. The header is absent on the last page.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Draft"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_DRAFT_CURSOR` when the cursor is malformed or is not a draft-list cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "listDrafts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of draft rows read for this page. Live access filtering can make the returned array shorter."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page. Read it from the Link header and send it unchanged."
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Literal draft search across sender, recipients, subject, and plain text."
          },
          {
            "in": "query",
            "name": "labelId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Require one assigned draft label."
          },
          {
            "in": "query",
            "name": "labelIds",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "maxLength": 100
              }
            },
            "style": "form",
            "explode": true,
            "description": "Repeat to require every selected draft label."
          }
        ]
      },
      "post": {
        "summary": "Create a draft",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Drafts"],
        "operationId": "createDraft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              },
              "example": {
                "mailboxId": null,
                "replyToMessageId": null,
                "forwardOfMessageId": null,
                "from": "",
                "to": [],
                "cc": [],
                "bcc": [],
                "subject": "",
                "text": "",
                "html": ""
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Audience-bound OAuth token. Request resource={origin}/api/v1.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "refreshUrl": "/api/auth/oauth2/token",
            "scopes": {
              "mail:read": "Read accessible mail",
              "mail:write": "Organize accessible mail",
              "mail:send": "Manage drafts and send mail",
              "offline_access": "Request a refresh token",
              "signatures:manage": "Manage signatures within the person's current access"
            }
          }
        }
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Same-origin Webmail web session. Secure deployments may prefix the cookie name."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "MailboxAddress": {
        "type": "object",
        "required": [
          "id",
          "mailboxId",
          "mailDomainId",
          "address",
          "displayName",
          "receiveEnabled",
          "sendEnabled",
          "isPrimary"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "mailboxId": {
            "type": "string"
          },
          "mailDomainId": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string",
            "description": "Shared sender name used with the mailbox address on future outbound messages."
          },
          "receiveEnabled": {
            "type": "boolean"
          },
          "sendEnabled": {
            "type": "boolean"
          },
          "isPrimary": {
            "type": "boolean"
          }
        }
      },
      "Mailbox": {
        "type": "object",
        "required": [
          "id",
          "address",
          "addresses",
          "displayName",
          "isActive",
          "accessLevel",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "format": "email"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailboxAddress"
            },
            "minItems": 1,
            "maxItems": 1
          },
          "displayName": {
            "type": "string",
            "description": "Shared sender name used with the mailbox address on future outbound messages."
          },
          "isActive": {
            "type": "boolean"
          },
          "accessLevel": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["read", "agent", "manager"]
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Label": {
        "type": "object",
        "required": ["id", "name", "color", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "color": {
            "type": "string",
            "enum": [
              "gray",
              "red",
              "orange",
              "amber",
              "green",
              "teal",
              "blue",
              "indigo",
              "purple",
              "pink"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ConversationPage": {
        "type": "object",
        "required": ["conversations", "nextCursor", "totalCount"],
        "properties": {
          "conversations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "totalCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "DraftInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DraftFields"
          },
          {
            "type": "object",
            "properties": {
              "signature": {
                "$ref": "#/components/schemas/SignatureSelection"
              }
            }
          }
        ]
      },
      "Draft": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DraftFields"
          },
          {
            "type": "object",
            "required": ["id", "version", "updatedAt", "attachments", "signature", "labels"],
            "properties": {
              "id": {
                "type": "string"
              },
              "version": {
                "type": "integer"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DraftAttachment"
                }
              },
              "signature": {
                "$ref": "#/components/schemas/SignatureSnapshot"
              },
              "labels": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Label"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}